🔥 Remove enable_location.sh script
This commit removes the `enable_location.sh` script from the
repository. The script was responsible for adding command-line
interfaces (CLIs) to the `clients.plist` file in the location
services directory on macOS systems.

The decision to remove this script was made to streamline the
codebase and eliminate unnecessary scripts that are no longer
required for the current functionality of the project.